+++ src/common/ast-path.js
@@ -16,8 +16,7 @@ class AstPath {
 
   /** @type {object} */
   get node() {
-    // eslint-disable-next-line unicorn/prefer-at -- `Array#at` is slow on Node.js v16 and v18
-    return this.stack[this.stack.length - 1];
+    return this.stack.at(-1);
   }
 
   /** @type {object | null} */