+++ src/index.js
@@ -4,15 +4,16 @@ import * as core from "./main/core.js";
 import { getSupportInfo as getSupportInfoWithoutPlugins } from "./main/support.js";
 import getFileInfoWithoutPlugins from "./common/get-file-info.js";
 import {
+  loadBuiltinPlugins,
   loadPlugins,
+  searchPlugins,
   clearCache as clearPluginCache,
-} from "./common/load-plugins.js";
+} from "./main/plugins/index.js";
 import {
   resolveConfig,
   resolveConfigFile,
   clearCache as clearConfigCache,
 } from "./config/resolve-config.js";
-import * as languages from "./languages.js";
 import * as errors from "./common/errors.js";
 import * as coreOptions from "./main/core-options.evaluate.js";
 import { createIsIgnoredFunction } from "./utils/ignore.js";
@@ -22,27 +23,33 @@ import arrayify from "./utils/arrayify.js";
 import partition from "./utils/partition.js";
 import isNonEmptyArray from "./utils/is-non-empty-array.js";
 
-const builtinPlugins = Object.values(languages);
-
 /**
  * @param {*} fn
- * @param {*} optsArgIdx
+ * @param {number} [optionsArgumentIndex]
  * @returns {*}
  */
 function withPlugins(
   fn,
-  optsArgIdx = 1 // Usually `opts` is the 2nd argument
+  optionsArgumentIndex = 1 // Usually `options` is the 2nd argument
 ) {
   return async (...args) => {
-    const opts = args[optsArgIdx] || {};
+    const options = args[optionsArgumentIndex] ?? {};
+    const { plugins = [], pluginSearchDirs } = options;
 
-    args[optsArgIdx] = {
-      ...opts,
-      plugins: [
-        ...builtinPlugins,
-        ...(await loadPlugins(opts.plugins, opts.pluginSearchDirs)),
-      ],
+    args[optionsArgumentIndex] = {
+      ...options,
+      plugins: (
+        await Promise.all([
+          loadBuiltinPlugins(),
+          // TODO: standalone version allow `plugins` to be `prettierPlugins` which is an object, should allow that too
+          loadPlugins(plugins),
+          options.pluginSearchDirs === false
+            ? []
+            : searchPlugins(pluginSearchDirs),
+        ])
+      ).flat(),
     };
+
     return fn(...args);
   };
 }