+++ src/language-html/embed.js
@@ -40,11 +40,14 @@ const cache = new WeakMap();
 function isVueSfcWithTypescriptScript(path, options) {
   const { root } = path;
   if (!cache.has(root)) {
-    const scriptLanguage = root.children.find((child) =>
-      isVueScriptTag(child, options)
-    )?.attrMap.lang;
-
-    cache.set(root, scriptLanguage === "ts" || scriptLanguage === "typescript");
+    cache.set(
+      root,
+      root.children.some(
+        (child) =>
+          isVueScriptTag(child, options) &&
+          ["ts", "typescript"].includes(child.attrMap.lang)
+      )
+    );
   }
 
   return cache.get(root);
@@ -73,7 +76,8 @@ async function printEmbeddedAttributeValue(path, htmlTextToDoc, options) {
       rootNode &&
       (rootNode.type === "ObjectExpression" ||
         rootNode.type === "ArrayExpression" ||
-        (options.parser === "__vue_expression" &&
+        ((options.parser === "__vue_expression" ||
+          options.parser === "__vue_ts_expression") &&
           (rootNode.type === "TemplateLiteral" ||
             rootNode.type === "StringLiteral")))
     ) {
@@ -121,11 +125,11 @@ async function printEmbeddedAttributeValue(path, htmlTextToDoc, options) {
 
   if (options.parser === "vue") {
     if (node.fullName === "v-for") {
-      return printVueFor(getValue(), attributeTextToDoc);
+      return printVueFor(getValue(), attributeTextToDoc, options);
     }
 
     if (isVueSlotAttribute(node) || isVueSfcBindingsAttribute(node, options)) {
-      return printVueBindings(getValue(), attributeTextToDoc);
+      return printVueBindings(getValue(), attributeTextToDoc, options);
     }
 
     /**
@@ -148,7 +152,9 @@ async function printEmbeddedAttributeValue(path, htmlTextToDoc, options) {
     if (isKeyMatched(vueEventBindingPatterns)) {
       const value = getValue();
       const parser = isVueEventBindingExpression(value)
-        ? "__js_expression"
+        ? isVueSfcWithTypescriptScript(path, options)
+          ? "__ts_expression"
+          : "__js_expression"
         : isVueSfcWithTypescriptScript(path, options)
         ? "__vue_ts_event_binding"
         : "__vue_event_binding";
@@ -157,13 +163,21 @@ async function printEmbeddedAttributeValue(path, htmlTextToDoc, options) {
 
     if (isKeyMatched(vueExpressionBindingPatterns)) {
       return printMaybeHug(
-        await attributeTextToDoc(getValue(), { parser: "__vue_expression" })
+        await attributeTextToDoc(getValue(), {
+          parser: options.__should_parse_vue_template_with_ts
+            ? "__vue_ts_expression"
+            : "__vue_expression",
+        })
       );
     }
 
     if (isKeyMatched(jsExpressionBindingPatterns)) {
       return printMaybeHug(
-        await attributeTextToDoc(getValue(), { parser: "__js_expression" })
+        await attributeTextToDoc(getValue(), {
+          parser: options.__should_parse_vue_template_with_ts
+            ? "__ts_expression"
+            : "__js_expression",
+        })
       );
     }
   }