+++ src/language-markdown/index.js
@@ -1,10 +1,14 @@
 import createParsers from "../utils/create-parsers.js";
-import { default as parsersConfig } from "./parsers.js";
 import printer from "./printer-markdown.js";
 
 export const printers = {
   mdast: printer,
 };
-export const parsers = createParsers(parsersConfig);
+export const parsers = createParsers([
+  {
+    importParsers: () => import("./parser-markdown.js"),
+    parserNames: ["remark", "markdown", "mdx"],
+  },
+]);
 export { default as languages } from "./languages.evaluate.js";
 export { default as options } from "./options.js";