+++ src/plugins/postcss.js
@@ -1,0 +1,2 @@
+export * from "../language-css/index.js";
+export * as default from "../language-css/index.js";