+++ src/plugins/typescript.js
@@ -1,0 +1,9 @@
+import createParsers from "../utils/create-parsers.js";
+
+export const parsers = createParsers([
+  {
+    importParsers: () => import("../language-js/parse/typescript.js"),
+    parserNames: ["typescript"],
+  },
+]);
+export default { parsers };