+++ src/standalone.js
@@ -1,26 +1,19 @@
 import * as core from "./main/core.js";
 import { getSupportInfo as getSupportInfoWithoutPlugins } from "./main/support.js";
-import * as prettierPluginJs from "./language-js/index.js";
-import * as prettierPluginJson from "./language-json/index.js";
-
-const builtinPlugins = [prettierPluginJs, prettierPluginJson];
 
 function withPlugins(
   fn,
-  optsArgIdx = 1 // Usually `opts` is the 2nd argument
+  optionsArgumentIndex = 1 // Usually `options` is the 2nd argument
 ) {
   // Returns Promises to consistent with functions in `index.js`
   // eslint-disable-next-line require-await
   return async (...args) => {
-    const opts = args[optsArgIdx] || {};
-    const plugins = opts.plugins || [];
+    const options = args[optionsArgumentIndex] ?? {};
+    const plugins = options.plugins ?? [];
 
-    args[optsArgIdx] = {
-      ...opts,
-      plugins: [
-        ...builtinPlugins,
-        ...(Array.isArray(plugins) ? plugins : Object.values(plugins)),
-      ],
+    args[optionsArgumentIndex] = {
+      ...options,
+      plugins: Array.isArray(plugins) ? plugins : Object.values(plugins),
     };
 
     return fn(...args);