+++ tests/integration/__tests__/third-party.js
@@ -1,8 +1,6 @@
 import path from "node:path";
 import { pathToFileURL } from "node:url";
 import createEsmUtils from "esm-utils";
-import outdent from "outdent";
-import stripAnsi from "strip-ansi";
 import { thirdParty } from "../env.js";
 import jestPathSerializer from "../path-serializer.js";
 
@@ -57,25 +55,6 @@ describe("cosmiconfig", () => {
       expect(filepath).toBe(file);
     });
   }
-
-  // #8815, please make sure this error contains code frame
-  test("Invalid json file", async () => {
-    const result = cosmiconfig("prettier").search(
-      path.join(__dirname, "../cli/config/invalid/broken-json")
-    );
-    await expect(result).rejects.toThrow();
-
-    try {
-      await result;
-    } catch (error) {
-      const message = stripAnsi(error.message);
-      expect(message).toContain(outdent`
-        > 1 | {a':}
-            |  ^
-          2 |
-      `);
-    }
-  });
 });
 
 test("isCI", async () => {